feat: add Tabulali project
Tabulali is a simple wallet aggregator that combines the data of multiple wallets
without the need to enter your seed phrase. The main use case is to keep track of
your portfolio in a convenient way.
It can be used as a website or even installed as a progressive web app on your phone.